feat: migrate models to ts
