+++ package.json
@@ -56,6 +56,7 @@
   },
   "devDependencies": {
     "@types/node": "^16.11.12",
+    "@types/uuid": "^8.3.3",
     "@typescript-eslint/eslint-plugin": "^5.6.0",
     "@typescript-eslint/parser": "^5.6.0",
     "chai": "4.3.4",
