+++ src/resources/columns/column.model.ts
@@ -1,6 +1,12 @@
 import { v4 as uuidv4 } from 'uuid';
 
 class Column {
+  id: string;
+
+  title: string;
+
+  order: number;
+
   constructor({ id = uuidv4(), title = 'Column1', order = 0 } = {}) {
     this.id = id;
     this.title = title;
