+++ src/resources/tasks/task.model.ts
@@ -1,13 +1,27 @@
 import { v4 as uuidv4 } from 'uuid';
 
 class Task {
+  id: string;
+
+  title: string;
+
+  order: number;
+
+  description: string;
+
+  userId: string | null;
+
+  boardId: string | null;
+
+  columnId: string | null;
+
   constructor({
     id = uuidv4(),
     title = 'New Task',
     order = 1,
     description = 'lol kek cheburek',
     userId = null,
-    boardId,
+    boardId = null,
     columnId = null,
   } = {}) {
     this.id = id;
