+++ src/resources/users/user.memory.repository.mjs
@@ -1,5 +1,5 @@
 import { addToDB, getDataFromDb } from '../utils';
-import User from './user.model.mjs';
+import User from './user.model';
 
 const getAll = async () => {
   const parsedData = await getDataFromDb();
