+++ src/resources/users/user.model.ts
@@ -1,6 +1,14 @@
 import { v4 as uuidv4 } from 'uuid';
 
 class User {
+  id: string;
+
+  name: string;
+
+  login: string;
+
+  password: string;
+
   constructor({
     id = uuidv4(),
     name = 'USER',
@@ -13,7 +21,7 @@ class User {
     this.password = password;
   }
 
-  static toResponse(user) {
+  static toResponse(user: User) {
     const { id, name, login } = user;
     return { id, name, login };
   }
